Simulation single file
Added a "set" method for the Terrain class + caught minor bugs.

Working on the required parts of a Simulation object as a whole before
taking it apart into smaller files.
Right now the simulation is running including the compass biped events
(albeit not really since nothing extends the leg after it's retracted).