Simulation Poincare
Worked on calculating the numerical poincare map.

If EndCond = 2, after convergence the IC and period are stored inside
the simulation object. Then sim.Poincare() can be called.
Alternatively, IClimCyc and Period can be set externally when the limit
cycle and period are already known.

The EndCond = 1 type of simulation was implemented. Here the robot needs
to walk a number on steps on the final slope, i.e., after the
inclination stops changing, if that's the case.

SimOut was merged with sim.End to create sim.Out, a single variable that
holds all the statistics.