Small bug
Now that sim.Period includes the number of steps for a complete cycle
and the step time, I used sim.Period(1) instead of sim.Period in
Poincare.m

NOTE: The period calculated is for the last step only. If the gait is
multi-period, a new simulation should be ran to get the IC at each step
and the period of each step in a full cycle.