Genome + small code for Controller/Simulation
On Controller added some code to correctly turn off an event triggered
pulse if the CPG resets while it's on.
On Simulation, changed Sim.Out when the stop button is pressed.
Created new Genome class and started working on adding Keys and how to
Decode them.
Added "CheckGenome" to Genome, haven't worked on it yet.