More debugging
* Genome
Added seq2str function for displaying a sequence nicely

* MOOGA
Changed the way of dealing with large eigenvalues
*** THIS IS STILL A BUG. Not sure how a controller can converge to a
period 1 LC and still have an eigenvalue larger than 1
The IC for checking the performance on up/downhill are now se to
IClimCyc if the controller converged.

* MOOGA/Run
Added some functions for showing the estimated time for the current
generation
Now the best genomes fitness is copied to the next generation so they
are not run again and again (big oopsie there)

* RunSeq
Added option of running the sequence for a time 'tend'

* GAfine
Implemented a function to increase the allowed genome range after each
generation

* RescueFit
Implemented a method for recovering the best genomes from all the
generations, in case some got lost along the way. I think I fixed the
bug that cause it so it might be unnecessary

* GAflat
Run this function for optimizing parameters only on flat terrain

* CompassBiped
Changed the "angle from vertical" condition that checks if the robot is
falling. Now it takes into account the incline as well.

* CompassBiped, Controller, Terrain
These classes now derive from mixin.Copyable so it's possible to
properly make copies of them

* Simulation
Implemented a deep copy method to copy also objects of Mod, Con and Env