Almost fully operational, still one bug to catch (at least)
* Controller
Removed deprecated pSoff (implementation for turning off pulses)
Added a reset function to set the pulses to 0 (needed when making a copy
of the Controller object)

* Mutate
Implemented a different way of dealing with bad mutation. There's still
a weird bug that appears only during the first generations and seems to
"break" all the genome sequences...

* InitGen
Corrected an old bug that shows up when using a small population as a
base for a new, larger population.

* MOOGA
Worked some more on the up/downhill fitness calculations and finally got
the simulations running properly! :+1:  (not the cleanest implementation
though)

* Simulation/Init
Corrected an initialization error thrown when Sim.IC is not initialized
and Graphics = 1.

* Terrain/Render
Vastly improved rendering speed by removing the for loop, now that
Terrain.Surf works with vectors :+1:            * Terrain: removed the
function handles which were causing trouble when copying the object and
moved back to the old implementation.