Most bugs caught, 99% functional
There are still some bugs out there. Mostly the one with the failed
sequences after the first (or second) generation. Also, it looks like
there was some problem with the ZMP calculation which still needs to be
checked.

* CompassBiped
Corrected the way in which it's checked whether the robot falls down
(large hip angle)

* Controller
Added an adaptation "low pass filter" to prevent strong jumps in torques
when the robot falls and gets a bad slope "measurement"

* GetTopPop
Added an option to use weighted population when selecting the best
genomes. Genomes from the first pareto front will have more chances of
reproducing than the ones on the second, third and so on.

* MOOGA/MOOGA
Added some more initialization parameters for up/downhill simulations.
Also, when removing the push-off torque from the ankle torques, the
pulse amplitude is not updated according to the slope (as happens in
simulation).

* MOOGA/Run
Added the required preparation for running the ZMPFit function and
implemented the weighted pairing

* RunSeq
Added the ZMPFit processing functions.

* Simulation/Render
Corrected the torque display bug cause when running two different
simulations on the same render figure

* GAfine
The mutation strength is now reduced after each generation (to converge
more easily)