More analysis and bug catching
* Analyze
Solved a bug caused when only a 0 slope is loaded from a saved file.
Moved most of the code to 2 functions: "WalkOnSlope" and
"SavePerformance".
Now 2 attempts are made for each slope, starting from the previous slope
and starting from a 0 slope. Also, 3 more slopes are tried after the
first failed one (so far this hasn't helped to find more stable slopes).
Also, slopes are now increased by base_d instead of 1, in preparation
for finer analysis.

* MOOGA/Find
Added the possibility of searching for fit genomes on previous
generations.

* MOOGA/Run -> Simulation/Render
Moved the cla command for clearing previous renders

* Simulation/Render
Changed the way in which the figure size/position is changed so it
doesn't reset when the user modifies it, but still works the first time.

* Simulation/Run
Caught a bug caused when the "Check clearance" statement is triggered (a
ground contact trying to slip through the cracks) and the impulsive
controller was not activated.

* Simulation/Simulation
Caught a bug when separating the impulsive torques, where lastPhi needed
to be updated before the first call to Controller.Adaptation.
Also improved the way in which impulsive torques are detected and
separated (MAY NEED MORE TESTING).