Poincare+ZMP+more bugs and new MOOGA options
* Controller
The Reset method can now reset the switches to a certain CPG phase. This
caused some one of the bugs with calculating the Poincare map.
Also, the adaptation function now receives the slope directly instead of
calculating it from the leg angles. This should also fix another bug
with calculating ZMP since the slope would change when the IC where
disturbed.

* Analyze
Added some formatting to the results to show "OK" or "FAILED" in green
and red after each run. Also applied the controller adaptation changes.

* MOOGA
Added an option to add new random genes to each generation.
Added a case to the Find method that shows the max fitness when called
without arguments and improved the way that output is displayed.

* MOOGA/Plot
Added some colors and moved the strsplit method to a separate file.

* MOOGA/Run
Corrected a bug with the detection of which fitness is being calculated
(uses strfind now to look for specific names).
Added post-processing for ZMPFit to "normalize" the fitness by
multiplying it by the largest slope attained (up or down).

* Simulation/Run
In order to process the Torques after running the different simulations,
a new variable "Slopes" was added to store the current Slope at each
moment.

* Simulation
Worked on improving the GetCleanTorques method.

* GAfine
Modified the code so that it can run with zero ankle torques
(under-actuated).