New slope fitness functions and some bug zapping
Commented out some of the "debugging" messages that I've been ignoring
for the past week anyway.

* Analyze
Corrected some bugs and added a code to skip the second attempt when the
first ran out of time (didn't fall or converge).
Moved most of the WalkOnSlope function to a method in Simulation.

* MOOGA - Fitness functions
Distance travelled is now calculated based on Sim.Out.SuppPos for all
fitness functions (except NrgEffFit).
Added a SlopeFit (Up/Down) function that replaces the old Up/Downhill
function. In this case the slope is increased gradually as well but
after each increase the robot is required to "show convergence" before
increasing the slope further.

* CheckConvergence
Added some code to check sooner if the simulation is converging (rather
than waiting until diff<minDiff). With "doGoNoGo" set to 1, the
simulation can be extended while converging to allow it to fully
converge (up to 3 times the original simulation time). With "doGoNoGo"
set to 2, the simulation can be stopped ahead of time if it shows that
it's converging/diverging.

* Simulation/Run
Commented out the "double-check" for ground clearance which seemed to be
causing more troubles than it solved.
T is now stored as sim.Out.T during the run (so it can be used by
CheckConvergence).

* Simulation
Added a JoinOuts method that allows to join the outputs from different
simulations (useful for ZMPFit and SlopeFit).
Updated GetCleanPulses to work properly with the new JoinOuts method.

* Terrain
Corrected the display bug in Terrain.