Improved Quantile and new Mimicking implementation
* .gitignore
Added some .mat files

* Controller
Added CompareSigs method to compare between the signals of 2 controllers
(should be the same length). The comparison is calculated as the
integral of abs(Sig1-Sig2) over T.

* Controller/MimicGains
Updates the controller's gains to mimic the torque signals of NCorig at
a certain slope. First the frequency is copied and then the amplitudes
are adjusted (this can be done with several methods but none is 100%
effective since the offset and duration of each pulse are not modified).

* Genome/Mimic
Uses Con.MimicGains so Seq1 mimics the output of Seq2 at a certain
slope. NOT FULLY IMPLEMENTED YET.

* Analyze
Added some post-analysis code to sort out the slopes+data (ascending
from negative to positive slopes).

* Pareto
Added an option to request a single front (or n fronts) to reduce
computations when not all fronts are needed.

* Quantile
Added a new method to remove the lower quantile that seems to work
nicely when the data is "sparse": delete IDs that scored 0 in several
categories (but save the first Pareto front of them, thus preserving the
maxima). From the last group (those that scored 0 in m categories and
would surpass the number of IDs that need to be discarded) we remove IDs
from Pareto fronts until just enough IDs are left in the group.

* GA_ROBIO
Reduced the mutation deltas (init and final) to see if fitness values
increase earlier (early generations).
(ALSO tried using "inverse" Pareto fronts but that didn't seem to work
well)