"Merging" ZMP and UP/Down slope fitness
Made some implementation changes to allow merging the ZMP fitness and
the up/down slope fitness, while still keeping each datum separately.

* GetTopPop
Now the best population is selected considering only the fitness values
in GA.FitIDs.

* InitGen
Added the modifications to GA.FitFcn (which now includes the name of the
function and the IDs on the Fit vector that it will occupy). MOST LIKELY
WILL NEED FURTHER IMPLEMENTATION, ESPECIALLY FOR BACKWARDS COMPATIBILITY
WHEN INPUTTING AN OLD FILE.

* MOOGA
Added the modifications to GA.FitFcn.
Modified SlopeFit to save all the output when walking on the first slope
(only the steps on the desired final slope are saved usually to reduce
computations).
Implemented the combined Up/Down+ZMP slope fitness function UDZMPFit.

* MOOGA/Plot
Added the modifications to GA.FitFcn.

* MOOGA/Run
Added the modifications to GA.FitFcn.

* RunSeq
Added the modifications to GA.FitFcn.