Bug squashing of new implementation
* MOOGA
INCREASED MaxBack to 0.2 (i.e. the max ankle length for scoring ZMP
points is 20cm)
Corrected the implementation of the new FitInd to properly display the
top results.

* Plot
Corrected the implementation of the new FitInd to properly display the
results.

* Quantile
Removed the option of duplicating the few top results when so many
scored 0 in almost all categories. This feature conflicted with the
Pareto implementation.

* MOOGA/Run
Added a call to the Neuron fire event when initializing the simulation.
Now the init cond for the CPG's phase is 0.

* RunSeq
Commented the call to handle a ground collision event when initializing
(only neuron fire event is called).

* Simulation/Run
That sneaky MOOGA was at it again... it managed to find a simulation
that sneaked under the ground and somehow ran underground super fast...
So, now we routinely check for "tunnels".

* WalkOnSlope
Instead of resetting the model stance foot back to (0,0), the terrain's
start at the last stance foot position.

* Terrain/Set
Added start_y and end_y as options to set. Don't know why they were
missing.