New ZMP fit limits and some improvements
* MOOGA
Fixed the implementation of Find so it can work with the new multiple
fitness output implementation.
Changed the ZMPFit output so it will reach 2 when the required foot
length is 0, 1 when the foot is as long as allowed and go quickly gown
to 0 as the foot gets longer.

* Plot
Added some code to normalize fitness values that go above 1 but still
display the maximum achieved.