Some more bugs and different MOOGA attempts
* Analyze
Reduced the step size.
Corrected the bug where the sorted slopes data wasn't saved.

* MOOGA
Added a SetFittest method to easily select amounts of top individuals to
keep / mutate and new random individuals as percentage of the population
(instead of fixed numbers).

* MOOGA/Plot
Corrected the fitness normalization code.

* MOOGA/Run
Commented the "post-processing" for ZMP which multiplied it by the best
up/down fitness achieved.

* Poincare
The simulation is now run without a set tstep (ode45 should choose
automatically) to prevent errors due to large tstep. To run faster, the
simulations are run without graphics when calculating the eigenvalues.

* Simulation/Render
Added an RSkip property to skip rendering some frames when the tstep is
very small. This can result in "jumpy" animations but it runs faster.

* Simulation/Run
Implemented the ability to run without a set tstep (it's set to 0.0111
to avoid conflicts with other parts of the code).

* Simulation
Increased the tolerance required for convergence (1e7 instead of 8e7)
and reduced the error size (from 1e4 to 1e5). Also implemented the
ability to run without a set tstep (it's set to 0.0111 to avoid
conflicts with other parts of the code).

Added some GA implementations to optimize controllers first on flat
terrain, then optimize FB gains on slopes. Also GenFBtuner runs a
simulation with a "flat genome" and allows to hand tune the FB gains.