More MOOGA improvements
Reduced the max front and max back size allowed for ZMP.
Limited the slope fitness to 15 degrees and updated the UDZMP fitness
to: (Slope+5*ZMP)/20

* WalkOnSlope
Corrected an issue where the robot would start walking "on air" when a
new slope was introduced (initial terrain height was wrong). This should
improve the number of genomes able to climb slopes larger than 2
degrees.

* GA_ROBIO
Added arguments for calling the function. This should help implement a
MOOGA that runs first with a large population and then with a small
population.

* ParetoPlot
Implemented a simple script for plotting 2D Pareto fronts.