ZMP implementation changes
Implemented the option of stopping the simulation if a ZMP value is
exceeded. The limits are defined by the foot size in CompassBiped (A2T
and A2H).

* CompassBiped
Added the A2T and A2H properties and included them in Set.
Implemented a GetZMP method which calculates the ZMP and returns the
min/max value. While coding it I discovered that GetGRF's implementation
was wrong since CB.Torques wasn't getting updated!

* GetCleanPulses
Modified the arguments required while keeping it backwards compatible

* Simulation/Run
Added a ZMP check after each simulation segment. If the limits are
exceeded the simulation stops.

* GA_ZMPcond
Added a new GA function which uses the new ZMP limited simulation.

STILL NEEDS TO BE IMPLEMENTED ON Analyze.m

--------------

Also
* MOOGA/Plot
Implemented an option to plot only the max fitness without averages

* Stages
Calls GA functions in stages.