Plots and small bugs
** Worked on compass biped model equations and calculated a ZMP bound

* Analyze
Added the zone calculations and eigenvalue sorting to the analyze method
so they are already included in the genome analysis .mat file. Also
moved the ConnectVectors and SortLines functions into Analyze.

* DisplayGen
Removed ConnectVector and SortLines. The code for loading the results is
now executed in Analyze (check if file exists and/or perform analysis).

* MOOGA/Plot & PlotGen
Added "Gen" for plotting genome results for the paper. Partially
implemented.

* RunSeq
Changed the call to the CPG fire event for the more general
Con.Reset(phase).

* GA_ZMPcond
Switched positions of gain genes so all the upslope genes are together
and all the downslope genes are together.

* ParetoPlot
Updated the script to plot the newest results with vel/energy and
up/downslope.

* Plots4Paper
Added a function for preparing plots for the CB paper. Just add a genome
results and hit Run.