Some more debugging and GA attempts
* Controller
Added a PhaseDiff method to return the distance between 2 phases when
the range is wrapped 1->0.

* InitGen
Added extra implementation for not copying the fitness results when ReDo
is set to 1.

* MOOGA
Implemented a method for combining the values of 2 previously calculated
fitness functions. In retrospect, this can be done with another fitness
function so I'll deprecate this and implement the other method.
Also, reduced the max slope fitness to 10 degrees since most genomes
don't make it to 10 anyway.

* MOOGA/Run
Implemented the fitness combination method described above.

* CheckConvergence
Used the new Controller PhaseDiff method to calculate delta_{k+1}.

* Poincare
Added a verification that resets the CPG phase if the disturbance sends
it past P_threshold.
Used the new Controller PhaseDiff method to calculate delta_{k+1}.
NOTE: This didn't solve the issue with jumping eigenvalues when a pulse
occurs just before/after ground impact. I believe that the issue is with
ode45 and events.

* GA_UpDown
New GA file to be implemented

* GA_ZMPcond
Added the combined fitness function described above. This helps produce
genomes that are good for walking up and downslopes.
NOTE: This