Convergence
* MOOGA
Only genomes that converged will get a score on UpDownFit (this should
weed out results that didn't converge).
Reduced the threshold required for Go-no-go convergence.

* CheckConvergence
Now the calculated diff gets stored in a Sim property.
Also changed Period to a single result (find 1 first) to reduce
computations.

* Simulation/Render
Now that ICdiff is stored, it can show the smallest diff and a guess as
to which period the gait will be.

* Plots4Paper
Changed the x axis of the torque signal to the CPG phase (instead of
time).