More preparations for slope-specific MOGA
* MOOGA\GetTopPop
* MOOGA\MOOGA
Implemented the "crowd control" algorithm to spread the results along
the pareto front. Genomes with the most different fitness within the top
pareto fronts are encouraged.
Still needs to be tested out.

* MOOGA\Run
Implemented an option to update the ICs stored in a genome using the
results from its simulation.

* Simulation\Run
* Simulation\UpdateStats
Removed a step taken when the last step causes the robot to fail.

* CrowdTest
Implemented the best method from MOGAcrowd to visualize what it selects
from the pareto front.

* GA_Slope
FileOut now contains the slope in its name.
Corrected a bug in Range.
Corrected a bug where the slope in IC was not in radians.
Currently running without RobustFit.

* MOGAcrowd
Implemented distance calculation method from NSGA-II algorithm.