First testing of slope specific MOGA
* GA_Slope
Changed the way in which the terrain is set because it caused a bug
where "end_slope" wasn't set, thus ruining the "walk n steps on a slope"
option of Sim.

* Terrain\Set
Added an error that is thrown when only start_slope or end_slope are set
with the "set" method. This is a quick patch but setting any start_* or
end_* alone will cause trouble if the terrain is a parabola.

* CompassBiped\CompassBiped
Prepared the event of swing leg switching direction (to extend the leg
with a reflex)

* Simulation\Run
Added the leg extension reflex, caused by event 3 of CompassBiped.

* MOOGA\GetTopPop
Made sure that we don't delete more genomes than necessary, i.e. that we
don't end up with length(TopIDs)<N

* MOOGA\MOOGA
Increased the length required to consider the energy fitness (so
controllers that take only one big step won't get a favorable fitness)

* MOOGA\Run
Corrected the Sim initialization so it's the same as RunSeq, i.e. the
CPG is reset to the IC phase and the swing leg is retracted (this could
be improved to decide whether to retract or not).
Also corrected the re-encoding of ICs back into the genome, where alpha
was being re-encoding as 2alpha.

* MOOGA\RunSeq
Removed the feedback for this simulation.