GA_Slope results!
* Controller\Controller
Implemented FBType = 3. It uses a look-up table of slopes to update the
frequency, amplitudes, offsets and durations of the controller.
Added "phase" as an input to method "Adaptation" so it's possible to
"reset" the updated controller's pulses to their on/off state at the
given CPG phase.

* MOOGA\MOOGA
Increased the requirement to get a non-zero score on energy fitness.

* Simulation\Init
* Simulation\Render
Decreased the window size so the simulation's graphics run faster.

* Terrain\Terrain
Added an option to have the sinusoidal slope "ride" on an incline plane
(incline+sine).

* GA_MultiSlope
Created a simple script to run the GA_Slope algorithm on different
slopes, making use of the results from previous runs.

* GA_Slope
Prepared the function to be called by GA_MultiSlope.

* SlopePlots
Collects the data obtained from running GA_MultiSlope and displays the
results.

* SlopeSim
CB simulation on an incline plane + sine to test the look-up table type
controller.