First version of Velocity GA
* CompassBiped
Made the leg extend when it reaches its apex (starts to fall to the
ground).

* Controller
Implemented the higher level signal "s_in" that tweaks the controller's
frequency and torques in a similar way as the slope feedback.
Removed timed leg extension.

* MOOGA
Added the Velocity Range Fitness function. It runs a simulation with the
CB walking on flat terrain and it increases/decreases s_in to make the
CB walk faster/slower.

* Simulation
Added some buttons to test the high-level signal s_in. In the future
they could be used to tweak any simulation parameter.