Matsuoka genome
Worked on encoding a Matsuoka oscillator based controller with a genome.

* Decode
Added a 'key' in the genome that tells the decoder to set the initial
conditions for the simulation so the neurons converge quickly to their
LC. It seems that setting 1 neuron is enough.

* Matsuoka class
Deleted unnecessary files like "AddPulse" and "MimicGains" that were
copied from Controller

* Matsuoka/Matsuoka
Prepared the class so it's easy to implement a controller with several
oscillators (2 half-neuron pairs for flexor/extensor = 4 DOF per
oscillator).

All the neurons get the same tau, tav, beta and a standard u0 of 1.

The inhibition between E/F neuron pairs can be set as single (equal for
all), pair specific or one for each connection.

The extra-neuronal inhibition is set from each extensor to each extensor
neuron and from each flexor to each flexor neuron.

The output matrix is set by dictating how many pulses will be used on
each joint, using the SetOutMatrix method.

* TestGen
Prepared a testing script for the Matsuoka Genome. The CB walks and
converges to a LC.

TO-DO: Adapt the MOOGA fitness functions to work with a matsuoka
controller.