Working GA with matsuoka based controller
* GA_Vel_Matsuoka
Prepared a version of the GA_Vel algorithm for the matsuoka based
controller. The algorithm searches for neuron parameters (tau, tav,
beta), connection weights (win, wex) and torque amplitudes (amp).

Only tested for walking on flat terrain with constant speed input.

* Decode
Updated the matsuoka IC "autoset" to change u0 rather than the IC. This
way the neuron's output should be normalized to 1 (+/-).

* Matsuoka
Added a saturation function to prevent the controller from generating
torques outside the given range.

Prepared the ground for high-level velocity input "s_in".

* MOOGA, Simulation
Moved the COT calculation from NrgFit to Simulation. Also modified the
calculation so it's easy to input different "gains" for
positive/negative work and changes in potential energy.

* CheckConvergence
"Fixed" a bug were Period would get set but not Increasing.

* Simulation rendering
Made a few cosmetic changes to the simulation rendering. Fixed an error
that occurred when the figure was closed (recursive calling of Render).
Changed the "Osc." status displayed to the output of GetPhasePerc to
allow for different implementations.