Working GA for Matsuoka with velocity range
Added a name property for each controller class so they're easy to
identify.

* MOOGA
Added a "FitMinMax" property to be able to distinguish between values
that we want to maximize and values that we want to minimize. (the
Pareto method still works only for maximization)

Corrected the way in which the average velocity was being calculated. It
seems that Mod.curSpeed had some glitch.

Added COT calculation to the DeltaVelFit method and corrected some bugs.

* Simulation/GetPeriod
Moved the period calculation to a separate method so it's easier to
calculate the period for Matsuoka oscillators.