Neuron state as output for Matsuoka
Initial work on implementing interconnected Matsuoka oscillators with
varying amplitude, where the amplitude is used directly as the neuron's
output.

* Decode
Set the IC value for the oscillators so it changes based on the
oscillator's parameters.

* Matsuoka
Added a W weight matrix that is the sum of the intra and extra-neuronal
connections. This matrix is multiplied and divided by each neuron's
desired amplitude so the connection weight is defined by Win / Wex and
not the relative amplitude of the neurons.

* Set
When setting the extra-neuronal connections, a weight is required for
each extensor and each flexor neuron (before it was only
extensor-extensor and flexor-flexor connections).