RunSeq + More changes to VGAM
* Genome
Added the right code to "change direction" for a Matsuoka genome (switch
between extensor and flexor to invert the torques).

* Matsuoka
Added an option to output 0 torque until a certain time, to allow the
neurons to converge to the LC first.

* MOOGA
Some bug corrections for the vel. range fitness.

* MOOGA/Run + RunSeq
Moved the code for running a genome simulation to RunSeq. Now, RunSeq is
called within the parfor loop.

* CheckConvergence
Implemented a new way to check go-no-go convergence based on stored
differences in initial conditions at the time of impact.

* Simulation/Run
For some reason a ground contact event was getting generated even if the
leg was not in touch with the ground. Added a verification to prevent
the CB from "jumping" into the air.