Updated MatsuokaML and debugging
* CheckGenome
Switched <, > for <= and => so the genome can be "bounded" with
min(max())

* Matsuoka / Derivative
Corrected sign for W matrix to use same equations as Matsuoka papers.

* Matsuoka / Set
Small correction to 'weights' option

* getConverged
Now returns the calculated period. Signals with "double period" are
marked as correct now and the max period is used as period.

* MatsuokaML
Added initialization that uses a pre-saved genome and includes a Sim
object.

* plotSamples
Some corrections to improve the plot's range

* processResults
Improved the period calculation and did extensive testing. Most signals
are now properly handled.

* processResultsFFT
Added Rea's version of period processing. Needs more work.

* runRandomSim
Implemented the use of genetic sequences to run random simulations.

* runScaledSim
Added a function similar to runRandomSim to run within parfor loop. It
re-scales tau_r, runs the sim again and stores the results.

* GA_Vel_Matsuoka
Updated the genome parameters.