Final simulations for paper
Some minor changes to CheckGenome and Decode to make sure that we know
when a random sequence is used.

Added a "GetGenes" method to extract selected genes from a given genetic
sequence. Useful for feeding a NN.

Added some improvements to the period calculation to try and get as many
signal periods right as possible. Some errors still result when
oscillations appear for certain ICs.
* getConverged
Fixed the period_diff calculation to catch signals with period multiples
(e.g. 1.6 and 3.2).

Updated the NN functions to return only tau_r (tau_a is calculated with
the ratio). NN trained using only the weights performed just as well as
those with weights+tonic input.

* MOOGA
Added a "Parents" property to store the ancestry of each individual.