Final final updates to generate paper results
Finally ignored Mats*.MAT and decided to silence the "genome out of
bounds" warnings.

* Matsuoka
Added option to set the tau_ratio to test if having a variable one
improves the NN results. It didn't.

* prepareNNData
Implemented method to normalize gene values for coupling weights to see
if it improves the NN performance. It didn't.

* trainNN
Normalized the performance errors.

* RunMatsuoka...
Tried different features/targets combinations for NN. Best performance
is still around 80%.