Matsuoka wrap-up and correcting bugs in MOGA
* Matsuoka
Corrected adaptation bug: Added tau0, tav0 parameters and made sure the
new parameters are bounded.

* MOOGA/GetDistance
Implemented a method to calculate the mean distance from each genome to
its closest neighbor. There probably is a faster implementation.

* MOOGA
Corrected the velocity range fitness: reduced the required velocity
change to 0.1%. Now the algorithm continues is the current sim
converged, was converging, or ran out of time. Also made it possible to
calculate this fitness when the original simulation didn't converge.

* Simulation
Added some code to verify that the torque is not constant (or zero)
throughout a step.