Finally good GA results
* MOOGA
Made some changes to VelRangeFit in order to better detect changes in
speed and to limit the run-time for each trial.

* MOOGA/Run
Added a simple loop to check outliers after running a simulation. It
should prevent "one-hit-wonders" (success that can't be reproduced) from
polluting the gene-pool.

* GA_Vel_Matsuoka
Corrected the way in which the NN's sequence is checked (it was keeping
the original sequence when the check failed, unlike with time RS)