Improvements to MML and preparations for GA of our CPG
* Controller
Added an option to set a single set of gains for faster (s>0) and slower
(s<0) inputs.

* Separated getNNPar into getNNin to be reused within MML
(prepareNNData) and MOGA. Also added option to normalize weights (by
amp).

* permSeq
Added a method to create permutated genetic sequences with the same
period. Permuting C and W or only W is optional.

* Added a simple method to show the parameter distributions at the end
of a MOGA run.

* Added a total runtime tracker for the MOGA to be stored with the GA
object.