12/02/2017 end day update
DID: finish to write the part of 4N symmetric CPG.
note: I added extra filtering of points. ignored samples that
ankle_period is vastly different than hip_period (or one of them is
NaN). also the new target period is the mean between the ankle and the
hip.

TODO: get more samples from 4N symmetric CPG because 10000 is not enough
for MoE (the perf is not greater than NN perf). probably there are not
enought points to go between the experts (maybe?).